service_type_generator/requirements.txt
pandas>=1.3.0
google-cloud-bigquery>=3.5.0
openpyxl>=3.0.10
